man.ponl: Reads main page from ponl.php, fixed formatting bug
The built-in script man now reads the main man page from ponl.php file.

Alos, a bug was discovered in readManPage() that can result in inproper
output of man page contents when an empty line is on the page. The * symbol
wasn't always removed in the output. This was fixed.

Signed-off-by: Jakob Rathbauer <ponl@software.jaraonthe.net>